requirements.txt
pymupdf
pdfplumber
pytesseract
langdetect
sentence-transformers
numpy
pdf2image
wordninja
indic-nlp-library

